Fixes bug with pat and look.
This fixes a bug that wouldn't let users enter a command containing pat or look if he/she wasn't standing next to a monster.

Now it will print a generic message for look if no monster is next to you.

Does not print anything if you are not next to a monster when useing pat.